Enum embedded_graphics_simulator::SimulatorEvent

pub enum SimulatorEvent {
    KeyUp {
        keycode: Keycode,
        keymod: Mod,
        repeat: bool,
    },
    KeyDown {
        keycode: Keycode,
        keymod: Mod,
        repeat: bool,
    },
    MouseButtonUp {
        mouse_btn: MouseButton,
        point: Point,
    },
    MouseButtonDown {
        mouse_btn: MouseButton,
        point: Point,
    },
    MouseWheel {
        scroll_delta: Point,
        direction: MouseWheelDirection,
    },
    MouseMove {
        point: Point,
    },
    Quit,
}

A derivation of sdl2::event::Event mapped to embedded-graphics coordinates

Variants

KeyUp

Fields

keycode: Keycode

The key being released

keymod: Mod

Any modifier being held at the time of keyup

repeat: bool

Whether the key is repeating

A keypress event, fired on keyUp

KeyDown

Fields

keycode: Keycode

The key being pressed

keymod: Mod

Any modifier being held at the time of keydown

repeat: bool

Whether the key is repeating

A keypress event, fired on keyDown

MouseButtonUp

Fields

mouse_btn: MouseButton

The mouse button being released

point: Point

The location of the mouse in Simulator coordinates

A mouse click event, fired on mouseUp

MouseButtonDown

Fields

mouse_btn: MouseButton

The mouse button being pressed

point: Point

The location of the mouse in Simulator coordinates

A mouse click event, fired on mouseDown

MouseWheel

Fields

scroll_delta: Point

The scroll wheel delta in the x and y direction

direction: MouseWheelDirection

The directionality of the scroll (normal or flipped)

A mouse wheel event

MouseMove

Fields

point: Point

The current mouse position

Mouse move event

Quit

An exit event