Struct embedded_graphics_simulator::SimulatorDisplay

pub struct SimulatorDisplay<C> { /* private fields */ }

Simulator display.

Implementations

impl<C: PixelColor> SimulatorDisplay<C>

pub fn with_default_color(size: Size, default_color: C) -> Self

Creates a new display filled with a color.

This constructor can be used if C doesn’t implement From<BinaryColor> or another default color is wanted.

pub fn get_pixel(&self, point: Point) -> C

Returns the color of the pixel at a point.

Panics

Panics if point is outside the display.

pub fn diff(
&self,
other: &SimulatorDisplay<C>
) -> Option<SimulatorDisplay<BinaryColor>>

Compares the content of this display with another display.

If both displays are equal None is returned, otherwise a difference image is returned. All pixels that are different will be filled with BinaryColor::On and all equal pixels with BinaryColor::Off.

Panics

Panics if the both display don’t have the same size.

impl<C> SimulatorDisplay<C>where
C: PixelColor + From<BinaryColor>,

pub fn new(size: Size) -> Self

Creates a new display.

The display is filled with C::from(BinaryColor::Off).

impl<C> SimulatorDisplay<C>where
C: PixelColor + Into<Rgb888>,

pub fn to_rgb_output_image(
&self,
output_settings: &OutputSettings
) -> OutputImage<Rgb888>

Converts the display contents into a RGB output image.

Examples

use embedded_graphics::{pixelcolor::Rgb888, prelude::*};
use embedded_graphics_simulator::{OutputSettingsBuilder, SimulatorDisplay};

let output_settings = OutputSettingsBuilder::new().scale(2).build();

let display = SimulatorDisplay::<Rgb888>::new(Size::new(128, 64));

// draw something to the display

let output_image = display.to_rgb_output_image(&output_settings);
assert_eq!(output_image.size(), Size::new(256, 128));

// use output image:
// example: output_image.save_png("out.png")?;

pub fn to_grayscale_output_image(
&self,
output_settings: &OutputSettings
) -> OutputImage<Gray8>

Converts the display contents into a grayscale output image.

Examples

use embedded_graphics::{pixelcolor::Gray8, prelude::*};
use embedded_graphics_simulator::{OutputSettingsBuilder, SimulatorDisplay};

let output_settings = OutputSettingsBuilder::new().scale(2).build();

let display = SimulatorDisplay::<Gray8>::new(Size::new(128, 64));

// draw something to the display

let output_image = display.to_grayscale_output_image(&output_settings);
assert_eq!(output_image.size(), Size::new(256, 128));

// use output image:
// example: output_image.save_png("out.png")?;

impl<C> SimulatorDisplay<C>where
C: PixelColor + ToBytes,
<C as ToBytes>::Bytes: AsRef<[u8]>,

pub fn to_be_bytes(&self) -> Vec<u8>

Converts the display content to big endian raw data.

pub fn to_le_bytes(&self) -> Vec<u8>

Converts the display content to little endian raw data.

pub fn to_ne_bytes(&self) -> Vec<u8>

Converts the display content to native endian raw data.

impl<C> SimulatorDisplay<C>where
C: PixelColor + From<Rgb888>,

pub fn load_png<P: AsRef<Path>>(path: P) -> ImageResult<Self>

Loads a PNG file.