Struct embedded_graphics_simulator::Window

pub struct Window { /* private fields */ }

Simulator window

Implementations

impl Window

pub fn new(title: &str, output_settings: &OutputSettings) -> Self

Creates a new simulator window.

pub fn update<C>(&mut self, display: &SimulatorDisplay<C>)where
C: PixelColor + Into<Rgb888> + From<Rgb888>,

Updates the window.

pub fn show_static<C>(&mut self, display: &SimulatorDisplay<C>)where
C: PixelColor + Into<Rgb888> + From<Rgb888>,

Shows a static display.

This methods updates the window once and loops until the simulator window is closed.

pub fn events(&mut self) -> impl Iterator<Item = SimulatorEvent> + '_

Returns an iterator of all captured SimulatorEvents.

Panics

Panics if called before update is called at least once.